requirements.txt
alembic==1.8.0
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.0.1
passlib==1.7.4
psycopg==3.0.8
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.19.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.15.0
watchgod==0.7
websockets==10.1